+++ sechecker/modules/find_netif_types.c
@@ -294,7 +294,7 @@ int find_netif_types_run(sechk_module_t * mod, apol_policy_t * policy, void *arg
 		qpol_context_t *if_con = NULL;
 		qpol_type_t *msg_type = NULL;
 		qpol_type_t *if_type = NULL;
-		int j = 0;
+		size_t j = 0;
 
 		netifcon = apol_vector_get_element(netifcon_vector, i);
 		qpol_netifcon_get_msg_con(q, netifcon, &msg_con);
@@ -383,7 +383,7 @@ int find_netif_types_print(sechk_module_t * mod, apol_policy_t * policy, void *a
 	unsigned char outformat = 0x00;
 	sechk_item_t *item = NULL;
 	sechk_proof_t *proof = NULL;
-	int i = 0, j = 0, k = 0, num_items = 0;
+	size_t i = 0, j = 0, k = 0, num_items = 0;
 	qpol_type_t *type;
 	qpol_policy_t *q = apol_policy_get_qpol(policy);
 	char *type_name;
